Task Bootstrap

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
@@ -35,8 +36,8 @@ class CommentController extends Controller
 
         $comment=new Comment();
         $comment->content=$valideted['content'];
-        $comment->id_user=0; // заглушка
-        $comment->id_post=$id_post; // заглушка
+        $comment->id_user=Auth::user()->id;
+        $comment->id_post=$id_post;
         $comment->save();
 
         return redirect('/post/'.$id_post.'/comments');
@@ -44,6 +45,10 @@ class CommentController extends Controller
 
     public function edit(string $id_post, string $id)
     {
+        $comment=Post::all()->where('id',$id)->first();
+        if(!Gate::allows('change-comment', $comment)){
+            return redirect('/post/'.$id_post)->with('message', 'Вы не можете редактировать чужой комментарий');
+        }
         return view('comment/comment_create',[
             'comment' => Comment::all()->where('id',$id)->first(),
             'id_post' => $id_post
@@ -65,6 +70,10 @@ class CommentController extends Controller
 
     public function destroy(string $id_post, string $id)
     {
+        $comment=Post::all()->where('id',$id)->first();
+        if(!Gate::allows('change-comment', $comment)){
+            return redirect('/post/'.$id_post)->with('message', 'Вы не можете удалить чужой комментарий');
+        }
         Comment::destroy($id);
         return redirect('/post/'.$id_post);
     }
